Create AccordionPanel component, simple form as example
This commit creates the accordion panel and accordion parent component.
Inside each AccordionPanel, I've also put a group of form inputs to show an example of what the accordion can be used for. This specific example shows a checkout form for shipping an order. One panel for contact info, and one panel for shipping info.